Update index.html
Moved the Basic Calulator project to the third spot and updated the
technology used on this project.

Added the LED Office Status Indicator project to the second place in the
work section. Added the description paragraph and included the technologies
used for this project. Project image will be added at a later time.